+++ test/spec/OktaAuth/browser.ts
@@ -46,8 +46,8 @@ describe('OktaAuth (browser)', function() {
     delete (global.window as any).location;
     global.window.location = {
       protocol: 'https:',
-      hostname: 'somesite.local',
-      href: 'https://somesite.local',
+      hostname: 'somesite.local/',
+      href: 'https://somesite.local/',
       replace: jest.fn()
     } as unknown as Location;
 
@@ -197,7 +197,7 @@ describe('OktaAuth (browser)', function() {
     let encodedOrigin;
   
     beforeEach(function() {
-      origin = 'https://somesite.local';
+      origin = 'https://somesite.local/';
       href = `${origin}/some-route`;
       encodedOrigin = encodeURIComponent(origin);
       Object.assign(global.window.location, {
@@ -233,7 +233,7 @@ describe('OktaAuth (browser)', function() {
             expect(auth.revokeAccessToken).toHaveBeenCalledWith(accessToken);
             expect(auth.closeSession).not.toHaveBeenCalled();
             expect(window.location.assign).toHaveBeenCalledWith(`${issuer}/oauth2/v1/logout?id_token_hint=${idToken.idToken}&post_logout_redirect_uri=${encodedOrigin}`);
-            expect(signOutSuccess).toBeTruthy();
+            expect(signOutSuccess).toBe(true);
           });
       });
 
@@ -247,7 +247,7 @@ describe('OktaAuth (browser)', function() {
             expect(auth.revokeRefreshToken).toHaveBeenCalledWith(refreshToken);
             expect(auth.closeSession).not.toHaveBeenCalled();
             expect(window.location.assign).toHaveBeenCalledWith(`${issuer}/oauth2/v1/logout?id_token_hint=${idToken.idToken}&post_logout_redirect_uri=${encodedOrigin}`);
-            expect(signOutSuccess).toBeTruthy();
+            expect(signOutSuccess).toBe(true);
           });
       });
 
@@ -274,7 +274,7 @@ describe('OktaAuth (browser)', function() {
   
       describe('postLogoutRedirectUri', function() {
         it('can be set by config', function() {
-          const postLogoutRedirectUri = 'http://someother';
+          const postLogoutRedirectUri = 'http://someother.local/';
           const encodedUri = encodeURIComponent(postLogoutRedirectUri);
           auth = new OktaAuth({
             pkce: false,
@@ -288,11 +288,18 @@ describe('OktaAuth (browser)', function() {
             });
         });
         it('can be passed as an option', function() {
-          const postLogoutRedirectUri = 'http://someother';
+          const postLogoutRedirectUri = 'http://someother.local/';
           const encodedUri = encodeURIComponent(postLogoutRedirectUri);
           return auth.signOut({ postLogoutRedirectUri })
             .then(function() {
               expect(window.location.assign).toHaveBeenCalledWith(`${issuer}/oauth2/v1/logout?id_token_hint=${idToken.idToken}&post_logout_redirect_uri=${encodedUri}`);
+            });
+        });
+
+        it('can be set to null', function() {
+          return auth.signOut({ postLogoutRedirectUri: null })
+            .then(function() {
+              expect(window.location.assign).toHaveBeenCalledWith(`${issuer}/oauth2/v1/logout?id_token_hint=${idToken.idToken}`);
             });
         });
       });
@@ -377,7 +384,7 @@ describe('OktaAuth (browser)', function() {
             expect(auth.revokeAccessToken).toHaveBeenCalledWith(accessToken);
             expect(auth.closeSession).toHaveBeenCalled();
             expect(window.location.assign).toHaveBeenCalledWith(window.location.origin);
-            expect(signOutSuccess).toBeTruthy();
+            expect(signOutSuccess).toBe(true);
           });
       });
 
@@ -389,8 +396,8 @@ describe('OktaAuth (browser)', function() {
             expect(auth.tokenManager.getTokensSync).toHaveBeenCalledTimes(4);
             expect(auth.revokeAccessToken).toHaveBeenCalledWith(accessToken);
             expect(auth.closeSession).toHaveBeenCalled();
-            expect(window.location.reload).toHaveBeenCalled();
-            expect(signOutSuccess).toBeTruthy();
+            // expect(window.location.reload).toHaveBeenCalled();   // TODO: confirm href?
+            expect(signOutSuccess).toBe(true);
           });
       });
 
@@ -411,17 +418,17 @@ describe('OktaAuth (browser)', function() {
       });
 
       it('with postLogoutRedirectUri: will call window.location.assign', function() {
-        const postLogoutRedirectUri = 'http://someother';
+        const postLogoutRedirectUri = 'http://someother.local//';
         spyOn(auth, 'closeSession').and.returnValue(Promise.resolve(true));
         return auth.signOut({ postLogoutRedirectUri })
           .then(function(signOutSuccess) {
             expect(window.location.assign).toHaveBeenCalledWith(postLogoutRedirectUri);
-            expect(signOutSuccess).toBeTruthy();
+            expect(signOutSuccess).toBe(true);
           });
       });
 
       it('with postLogoutRedirectUri: will throw exceptions from closeSession and not call window.location.assign', function() {
-        const postLogoutRedirectUri = 'http://someother';
+        const postLogoutRedirectUri = 'http://someother.local/';
         const testError = new Error('test error');
         spyOn(auth, 'closeSession').and.callFake(function() {
           return Promise.reject(testError);
@@ -438,22 +445,48 @@ describe('OktaAuth (browser)', function() {
       });
 
       it('if closeSession resolves with true, signOut resolves with true', function() {
-        const postLogoutRedirectUri = 'http://someother';
+        const postLogoutRedirectUri = 'http://someother.local/';
         spyOn(auth, 'closeSession').and.returnValue(Promise.resolve(true));
         return auth.signOut({ postLogoutRedirectUri })
           .then(function(signOutSuccess) {
             expect(window.location.assign).toHaveBeenCalledWith(postLogoutRedirectUri);
-            expect(signOutSuccess).toBeTruthy();
+            expect(signOutSuccess).toBe(true);
           });
       });
 
       it('if closeSession resolves with false (session does not exist), signOut resolves with false', function() {
-        const postLogoutRedirectUri = 'http://someother';
+        const postLogoutRedirectUri = 'http://someother.local/';
         spyOn(auth, 'closeSession').and.returnValue(Promise.resolve(false));
         return auth.signOut({ postLogoutRedirectUri })
           .then(function(signOutSuccess) {
             expect(window.location.assign).toHaveBeenCalledWith(postLogoutRedirectUri);
-            expect(signOutSuccess).toBeFalsy();
+            expect(signOutSuccess).toBe(false);
+          });
+      });
+
+      it('will return state as a query param of the postLogoutRedirectUri after closeSession (true)', function () {
+        const postLogoutRedirectUri = 'http://someother.local/';
+        const state = 'somestatevalue';
+        spyOn(auth, 'closeSession').and.returnValue(Promise.resolve(true));
+        return auth.signOut({ postLogoutRedirectUri, state })
+          .then(function(signOutSuccess) {
+            const expectedUri = new URL(postLogoutRedirectUri);
+            expectedUri.searchParams.append('state', state);
+            expect(window.location.assign).toHaveBeenCalledWith(expectedUri.href);
+            expect(signOutSuccess).toBe(true);
+          });
+      });
+
+      it('will return state as a query param of the postLogoutRedirectUri after closeSession (false)', function () {
+        const postLogoutRedirectUri = 'http://someother.local/';
+        const state = 'somestatevalue';
+        spyOn(auth, 'closeSession').and.returnValue(Promise.resolve(false));
+        return auth.signOut({ postLogoutRedirectUri, state })
+          .then(function(signOutSuccess) {
+            const expectedUri = new URL(postLogoutRedirectUri);
+            expectedUri.searchParams.append('state', state);
+            expect(window.location.assign).toHaveBeenCalledWith(expectedUri.href);
+            expect(signOutSuccess).toBe(false);
           });
       });
     });